+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+### versio 0.1.6
+ - `Inertia` enum deprecated. `Image360Controller.inertia` now is a `Float` value.
+ - `Image360Contoller` now have flag `isGestureControlEnabled` to enable/disable gesture control.
+ - Device motion control for `Image360Contoller` implemented. This feature could be enabled/disabled via new `isDeviceMotionControlEnabled` flag.
+
 ### version 0.1.5
  - `Image360View` now has special orientation subview. It's controlled via `isOrientationViewHidden` property.